+++ src/App.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import './App.css';
 import DeviceList from './components/DeviceList';
-import DeviceDataChart from './components/DeviceDataChart';
+import DeviceDataCharts from './components/DeviceDataCharts';
 
 function App() {
   return (
     <div>
       <DeviceList />
-      <DeviceDataChart />
+      <DeviceDataCharts />
     </div>
   );
 }
